Added colors for message origins and cleaned up javascript
The previous commit on this feature branch actually closes #6.  The javascript added in https://github.com/Rythious/ScratchMUD/commit/70fdf96649a82fb43a03b909f7a96df28c7db430 checks to see if the user is scrolled to the bottom.  If they are, it puts the scrollbar back to the bottom after adding new text.  This commit closes #10 as it adds colors to the different sources of messages.  Right now there is messages from clients, messages from the server, and a special color for room descriptions.