Updated the database and rescaffolded
I realized after trying to add two target dummy NPCs to a single room that my compound key in the database didn't make sense when I wanted an entity to have the possibility of having multiple copies of a thing.  Created identity columns for these tables and regenerated the models and context.